Smooth Probabilities

Calculate probabilities for multiple population. Given data on the number of 'trials' and 'succession' in each population, calculate the probability of success in each population. For instance, given data on the number of respondent, and number of employed respondents, by area, calculate the probability of being employed in each area. The probabilities are smoothed: the values are shifted towards the overall mean, with values that are based on small sample sizes being shifted the furtherst.

Usage

smooth_prob(x, size, prior_cases = 10)

Arguments

x

Number of successes in each population. A numeric vector.

size

Number of trials in each population. A numeric vector.

prior_cases

Parameter controlling smoothing. Default is 10.

Value

A numeric vector with the smoothed probabilities.

Stratifying

It is often appropriate to stratify a population and smooth separately within each strata. For instance, when estimating probabilities of being employed, it may be appropriate to divide the population into strata defined by age and sex.

The easiest way to do stratified smoothing is to use grouped data frames. See below for an example.

prior_counts

The argument prior_counts controls the degree of smoothing. It only has a noticeable effect when the number of areas, and the population per area, is small. Larger values for prior_counts produce more smoothing.

Mathematical details

The smoothing is based on the model

$$x_k \sim \text{Binom}(n_k, \pi_k)$$

$$\pi_k \sim \text{Beta}(\lambda \nu, (1 - \lambda) \nu)$$

$$\lambda \sim \text{Unif}(0, 1)$$

$$\nu \sim \text{LogNormal}^+(\log M, 1)$$

where

\(k\) indexes area or population,

\(x_k\) is the number of successes, which is specified by argument x

\(n_k\) is the number of trials, which is specified by argument size

\(\pi_k\) is the probability of success, and

\(M\) control smoothing, and can be specified by argument prior_counts.

smooth_prob() returns \(\hat{\pi}_k\), the maximum posterior density estimate of \(\pi_k\).

The "direct" (unsmoothed) estimate of the probability of success is \(x_k / n_k\).

For details on the model, see the vignette LINK TO VIGNETTE.

Examples

## use synthetic census data
census <- smoothscale::syn_census

## smooth all groups towards the national level
smoothed <- smooth_prob(x = census$child_labour,
                        size = census$all_children)
smoothed
#>   [1] 0.1205323 0.3007721 0.2486283 0.4217355 0.2356759 0.3776814 0.2540063
#>   [8] 0.2058644 0.2924516 0.2239079 0.2545633 0.1350277 0.2660161 0.2883652
#>  [15] 0.3437973 0.3675160 0.1483832 0.2450890 0.1086436 0.2215408 0.2854412
#>  [22] 0.3709231 0.2163642 0.1233066 0.1982336 0.2856093 0.2248202 0.3380457
#>  [29] 0.2318525 0.1311424 0.3165574 0.1768094 0.2755290 0.3578148 0.2317961
#>  [36] 0.3878768 0.2242419 0.1923907 0.2302324 0.2967493 0.2996808 0.4162655
#>  [43] 0.2406016 0.2450890 0.2121073 0.1626019 0.2694820 0.2510393 0.2432360
#>  [50] 0.2199017 0.1408026 0.3073848 0.2533793 0.4574859 0.2921075 0.2556513
#>  [57] 0.2635971 0.3298322 0.1651697 0.2448263 0.3084629 0.3046090 0.2546887
#>  [64] 0.4533667 0.3293704 0.4941331 0.1241034 0.1748751 0.1068550 0.2660161
#>  [71] 0.2473347 0.4071273 0.2089473 0.1378428 0.2476872 0.2579587 0.2519462
#>  [78] 0.2980855 0.2919887 0.1101094 0.3667327 0.2129209 0.3445193 0.3279501
#>  [85] 0.2258112 0.4045270 0.2161521 0.1724420 0.2668160 0.2101998 0.3169499
#>  [92] 0.3913688 0.1230403 0.3153029 0.2660161 0.2442866 0.2153822 0.3032731
#>  [99] 0.3212769 0.2769412 0.1804302 0.2918149 0.3369369 0.5227290 0.3764610
#> [106] 0.3776814 0.2635971 0.4442049 0.2071769 0.3283814 0.3625966 0.3724415
#> [113] 0.1626019 0.4463724 0.4037881 0.5615725 0.1091029 0.2450890 0.1543658
#> [120] 0.2783979 0.3006354 0.4927858 0.1758988 0.1678941 0.3575088 0.3519603
#> [127] 0.3019739 0.4611818 0.3249433 0.1603462 0.2715621 0.2129209 0.3852343
#> [134] 0.4221915 0.3932817 0.5226584 0.2912068 0.1923907 0.2975963 0.3881023
#> [141] 0.3507243 0.4380781 0.1344564 0.3153029 0.3398517 0.3131803 0.2153822
#> [148] 0.3307152 0.3590365 0.2715621 0.1784770 0.4173126 0.3736966 0.5401010
#> [155] 0.4517846 0.2043282 0.2635971 0.4750223 0.2516317 0.3082693 0.3771943
#> [162] 0.2706927 0.2215408 0.4739919 0.4463724 0.5390927 0.1483832 0.2540063
#> [169] 0.1794613 0.3714888 0.2700076 0.5301347 0.2574875 0.2312844 0.2812681
#> [176] 0.3702633 0.3419224 0.4440680 0.3714888 0.2036228 0.4239965 0.1862431
#> [183] 0.4093397 0.3940101 0.3322635 0.5083383 0.3290645 0.2888125 0.3442008
#> [190] 0.2017042 0.4047612 0.5542312 0.1145517 0.3046090 0.2318525 0.2972702
#> [197] 0.4096207 0.4148709 0.3357356 0.3235478

## compare smoothed and unsmoothed ("direct") estimates
unsmoothed <- census$child_labour / census$all_children
rbind(head(smoothed), head(unsmoothed))
#>           [,1]      [,2]      [,3]      [,4]      [,5]      [,6]
#> [1,] 0.1205323 0.3007721 0.2486283 0.4217355 0.2356759 0.3776814
#> [2,] 0.1141304 0.3030303 0.2472406 0.4265537 0.2277228 0.6000000

## use tidyverse functions to smooth
## each age-sex group towards a
## different average
library(dplyr, warn.conflicts = FALSE)
census |>
  group_by(age, sex) |>
  mutate(smoothed = smooth_prob(x = child_labour,
                                size = all_children))
#> # A tibble: 200 × 6
#> # Groups:   age, sex [4]
#>    area    age   sex    child_labour all_children smoothed
#>    <chr>   <chr> <chr>         <int>        <dbl>    <dbl>
#>  1 Area 01 5-9   Female           42          368    0.121
#>  2 Area 02 5-9   Female           10           33    0.278
#>  3 Area 03 5-9   Female          112          453    0.247
#>  4 Area 04 5-9   Female          151          354    0.416
#>  5 Area 05 5-9   Female           23          101    0.230
#>  6 Area 06 5-9   Female            3            5    0.308
#>  7 Area 07 5-9   Female            3           14    0.231
#>  8 Area 08 5-9   Female            6           35    0.199
#>  9 Area 09 5-9   Female           14           48    0.276
#> 10 Area 10 5-9   Female          422         1889    0.224
#> # ℹ 190 more rows